+++ VentasProyect/Models/t_venta.cs
@@ -25,8 +25,8 @@ namespace VentasProyect.Models
         public Nullable<int> usu_id { get; set; }
         public Nullable<System.DateTime> ven_fecha { get; set; }
         public string ven_metodo_pago { get; set; }
-        public Nullable<int> ven_total { get; set; }
-        public Nullable<int> ven_numero_transaccion { get; set; }
+        public Nullable<long> ven_total { get; set; }
+        public Nullable<long> ven_numero_transaccion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<t_detalle_venta> t_detalle_venta { get; set; }
